% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasteval-package.R
\docType{package}
\name{fasteval-package}
\alias{fasteval-package}
\title{fasteval: Fast evaluation of mathematical expressions.}
\description{
This package quickly converts strings like "1+2" to a value of 3, 
including variable interpolation. The only R function, 
\code{\link{fasteval}}, is a thin wrapper around the C library 
\code{{tinyexpr}}. It is generally 50-150 times faster than 
looping over a vector such as 
\code{lapply(x, function(x) eval(parse(text = x)))}.
}
\note{
Unlike the original C library, exponentiation precedence takes place from
right to left to match the behavior in R. Additionally, \code{"log"}
refers to the natural log, as it does in R.
}
\references{
Van Winkle, Lewis. C Library, \code{{tinyexpr}}. https://github.com/codeplea/tinyexpr
}
