% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{five_moves}
\alias{five_moves}
\title{Five arithmetic moves to combine 6 numbers}
\format{
A character vector with 463,164 elements:
\itemize{
  \item{\eqn{((((1*2)*4)-5)*7)-8}}
  \item{\eqn{((((1*2)*4)-5)*7)*8}}
  \item{\eqn{((((1*2)*4)-5)*7)+8}}
  \item{\eqn{...}}
  \item{\eqn{8/(7/((4+5)-(2/1)))}}
  \item{\eqn{8/(7/((5-(1*2))+4))}}
  \item{\eqn{8/(7/((5-(2/1))+4))}}
}
}
\usage{
five_moves
}
\description{
Given the numbers 1, 2, 4, 5, 7, 8; how can you combine them with 
  addition, subtraction, multiplication, and division, in any order,
  such that each intermediate step is always a positive integer?
  This dataset lists all such unique permutations, and the results can
  be quickly calculated.
}
\examples{
moves <- data.frame(string = five_moves)
moves$result <- fasteval(moves$string)
head(moves)
}
\keyword{datasets}
