% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{four_moves}
\alias{four_moves}
\title{Four arithmetic moves to combine 5 numbers}
\format{
A character vector with 25,563 elements:
\itemize{
  \item{\code{((((1*2)*4)-5)*7)}}
  \item{\code{((((1*2)*4)-5)+7)}}
  \item{\code{((((1*2)*4)-7)*5)}}
  \item{\code{...}}
}
}
\usage{
four_moves
}
\description{
Given the numbers 1, 2, 4, 5, 7; how can you combine them with 
  addition, subtraction, multiplication, and division, in any order,
  such that each intermediate step is always a positive integer?
  This dataset lists all such permutations, and the results can
  be quickly calculated.
}
\examples{
moves <- data.frame(string = four_moves)
moves$result <- fasteval(moves$string)
head(moves)
}
\keyword{datasets}
